requirements/requirements-dev.txt
-r requirements-test.txt

# Debugging
ipdb==0.10.0
ipython==4.2.0

# Packaging
bumpversion==0.5.3
wheel==0.29.0

# Style Enforcement
isort==4.2.5
pylint-django==0.7.1
pylint-common==0.2.2

# Test Runner
pytest-django==2.9.1
pytest-sugar==0.7.1
